:sparkles: Add support for text compression
Future Jenson files will be compressed using the Brotli compression
algorithm to reduce the file size on disk.

Here's how the file size compares with compression on the test files
in the testing suite:
- Uncompressed debugging file (marteau): 3161 bytes
- Uncompressed: 4346 bytes
- Compressed: 1348 bytes

The result is about 31% in savings, allowing more for more space in
apps and games.

Likewise, the times don't seem to be affected, either. Here's how
the performance compares with compression, using the unit tests
that measure the performance:

- Read (uncompressed): 0.000593s (avg)
- Write (uncompressed): 0.0014s (avg)
- Read (compressed): 0.000464s (avg)
- Write (compressed): 0.00197s (avg)

The read operations are just slightly faster with a 0.000129s diff.
The write operations are a bit slower, though the main game won't
be writing Jenson files, so this should be acceptable. The only case
where this may be a problem would be with Marteau.

Overall, the compression would be a good addition to the format,
allowing for games and apps to have more space for more essential
content.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>