:rocket: Prepare for Jenson v2
This commit makes several changes to both the Jenson format and the
internal implementations.

First, the Code of Conduct, SwiftFormat configuration, and SwiftLint
configuration files have been copied from Indexing Your Heart as
this package is governed under the organization.

Next, most of the structs included in this package now conform to
Identifiable, Equatable, and Hashable. This will prove useful for
SwiftUI apps that utilize these structs.

Finally, the package has been updated to support Jenson v2, which
includes a new refresh event, an optional application field for
third-party apps and/or APIs, and a story metadata field.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>